requirements.txt
aiogram
PyPDF2==1.26.0
pillow==8.3.1
requests==2.25.1
unidecode==1.3.2